+++ packages/app/src/domain/layout/layout.tsx
@@ -5,7 +5,6 @@ import { SEOHead } from '~/components/seoHead';
 import { AppFooter } from '~/components-styled/layout/app-footer';
 import { AppHeader } from '~/components-styled/layout/app-header';
 import { SkipLinkMenu } from '~/components-styled/skip-link-menu';
-import { ILastGeneratedData } from '~/static-props/last-generated-data';
 
 interface LayoutProps {
   title: string;
@@ -48,7 +47,7 @@ export function getLayout(layoutProps: LayoutProps, lastGenerated: string) {
 }
 
 export function Layout(
-  props: LayoutProps & ILastGeneratedData & { children: React.ReactNode }
+  props: LayoutProps & { lastGenerated: string; children: React.ReactNode }
 ) {
   const {
     children,
