+++ packages/app/src/pages/gemeente/[code]/index.tsx
@@ -1,10 +1,9 @@
 import { FCWithLayout } from '~/domain/layout/layout';
 import { getMunicipalityLayout } from '~/domain/layout/municipality-layout';
-import { getGmData, getLastGeneratedDate } from '~/static-props/data';
-import { getPaths } from '~/static-props/gm-data';
-import { createGetStaticProps } from '~/static-props/utils/create-get-static-props';
+import { getGmData, getLastGeneratedDate } from '~/static-props/get-data';
+import { createGetStaticProps } from '~/static-props/create-get-static-props';
 
-export const getStaticPaths = getPaths();
+export { getStaticPaths } from '~/static-paths/gm';
 
 export const getStaticProps = createGetStaticProps(
   getLastGeneratedDate,
