+++ packages/app/src/pages/gemeente/[code]/ziekenhuis-opnames.tsx
@@ -15,16 +15,15 @@ import { SEOHead } from '~/components/seoHead';
 import { FCWithLayout } from '~/domain/layout/layout';
 import { getMunicipalityLayout } from '~/domain/layout/municipality-layout';
 import siteText from '~/locale/index';
+import { createGetStaticProps } from '~/static-props/create-get-static-props';
 import {
   createGetChoroplethData,
   getGmData,
   getLastGeneratedDate,
-} from '~/static-props/data';
-import { getPaths } from '~/static-props/gm-data';
-import { createGetStaticProps } from '~/static-props/utils/create-get-static-props';
+} from '~/static-props/get-data';
 import { replaceVariablesInText } from '~/utils/replaceVariablesInText';
 
-const text = siteText.gemeente_ziekenhuisopnames_per_dag;
+export { getStaticPaths } from '~/static-paths/gm';
 
 export const getStaticProps = createGetStaticProps(
   getLastGeneratedDate,
@@ -33,6 +32,8 @@ export const getStaticProps = createGetStaticProps(
     gm: ({ hospital_nice }) => ({ hospital_nice }),
   })
 );
+
+const text = siteText.gemeente_ziekenhuisopnames_per_dag;
 
 const IntakeHospital: FCWithLayout<typeof getStaticProps> = (props) => {
   const { data, choropleth, municipalityName } = props;
@@ -137,7 +138,5 @@ const IntakeHospital: FCWithLayout<typeof getStaticProps> = (props) => {
 };
 
 IntakeHospital.getLayout = getMunicipalityLayout();
-
-export const getStaticPaths = getPaths();
 
 export default IntakeHospital;
