+++ packages/app/src/pages/landelijk/gedrag.tsx
@@ -18,8 +18,8 @@ import {
   createGetChoroplethData,
   getNlData,
   getLastGeneratedDate,
-} from '~/static-props/data';
-import { createGetStaticProps } from '~/static-props/utils/create-get-static-props';
+} from '~/static-props/get-data';
+import { createGetStaticProps } from '~/static-props/create-get-static-props';
 
 const text = siteText.nl_gedrag;
