+++ packages/app/src/pages/landelijk/maatregelen.tsx
@@ -10,8 +10,8 @@ import { SEOHead } from '~/components/seoHead';
 import { FCWithLayout } from '~/domain/layout/layout';
 import { getNationalLayout } from '~/domain/layout/national-layout';
 import text from '~/locale';
-import { getNlData, getLastGeneratedDate } from '~/static-props/data';
-import { createGetStaticProps } from '~/static-props/utils/create-get-static-props';
+import { getNlData, getLastGeneratedDate } from '~/static-props/get-data';
+import { createGetStaticProps } from '~/static-props/create-get-static-props';
 import theme from '~/style/theme';
 import { useEscalationLevel } from '~/utils/use-escalation-level';
