+++ packages/app/src/pages/landelijk/thuiswonende-ouderen.tsx
@@ -20,8 +20,8 @@ import {
   createGetChoroplethData,
   getNlData,
   getLastGeneratedDate,
-} from '~/static-props/data';
-import { createGetStaticProps } from '~/static-props/utils/create-get-static-props';
+} from '~/static-props/get-data';
+import { createGetStaticProps } from '~/static-props/create-get-static-props';
 
 const text = siteText.thuiswonende_ouderen;
