+++ packages/app/src/pages/landelijk/ziekenhuis-opnames.tsx
@@ -25,8 +25,8 @@ import {
   createGetChoroplethData,
   getNlData,
   getLastGeneratedDate,
-} from '~/static-props/data';
-import { createGetStaticProps } from '~/static-props/utils/create-get-static-props';
+} from '~/static-props/get-data';
+import { createGetStaticProps } from '~/static-props/create-get-static-props';
 
 const text = siteText.ziekenhuisopnames_per_dag;
