+++ packages/app/src/pages/veiligheidsregio/[code]/gehandicaptenzorg.tsx
@@ -12,23 +12,22 @@ import { SEOHead } from '~/components/seoHead';
 import { FCWithLayout } from '~/domain/layout/layout';
 import { getSafetyRegionLayout } from '~/domain/layout/safety-region-layout';
 import siteText from '~/locale/index';
-import { getVrData, getLastGeneratedDate } from '~/static-props/data';
-import { createGetStaticProps } from '~/static-props/utils/create-get-static-props';
-import { getPaths } from '~/static-props/vr-data';
+import { createGetStaticProps } from '~/static-props/create-get-static-props';
+import { getLastGeneratedDate, getVrData } from '~/static-props/get-data';
 import { replaceVariablesInText } from '~/utils/replaceVariablesInText';
+
+export { getStaticPaths } from '~/static-paths/vr';
+
+export const getStaticProps = createGetStaticProps(
+  getLastGeneratedDate,
+  getVrData
+);
 
 const locationsText =
   siteText.veiligheidsregio_gehandicaptenzorg_besmette_locaties;
 const positiveTestPeopleText =
   siteText.veiligheidsregio_gehandicaptenzorg_positief_geteste_personen;
 const mortalityText = siteText.veiligheidsregio_gehandicaptenzorg_oversterfte;
-
-export const getStaticPaths = getPaths();
-
-export const getStaticProps = createGetStaticProps(
-  getLastGeneratedDate,
-  getVrData
-);
 
 const DisabilityCare: FCWithLayout<typeof getStaticProps> = (props) => {
   const { data, safetyRegionName } = props;
