+++ packages/app/src/static-props/create-get-static-props.tsx
